+++ lib/geiger_api.dart
@@ -2,19 +2,19 @@
 
 library geiger_api;
 
-export 'src/communication/communication_api_factory.dart';
-export 'src/communication/communication_api.dart';
-export 'src/communication/communication_exception.dart';
-export 'src/communication/declaration.dart';
-export 'src/communication/plugin_information.dart';
-export 'src/communication/declaration_mismatch_exception.dart';
-export 'src/communication/geiger_api.dart';
-export 'src/communication/geiger_url.dart';
-export 'src/communication/menu_item.dart';
-export 'src/communication/message.dart';
-export 'src/communication/message_listener.dart';
-export 'src/communication/message_type.dart';
-export 'src/communication/plugin_listener.dart';
-export 'src/communication/plugin_starter.dart';
-export 'src/communication/storable_hash_map.dart';
-export 'src/communication/storable_string.dart';
+export 'src/communication_api_factory.dart';
+export 'src/communication_api.dart';
+export 'src/exceptions/communication_exception.dart';
+export 'src/plugin/declaration.dart';
+export 'src/plugin/plugin_information.dart';
+export 'src/exceptions/declaration_mismatch_exception.dart';
+export 'src/geiger_api.dart';
+export 'src/message/geiger_url.dart';
+export 'src/plugin/menu_item.dart';
+export 'src/message/message.dart';
+export 'src/message/message_listener.dart';
+export 'src/message/message_type.dart';
+export 'src/plugin/plugin_listener.dart';
+export 'src/plugin/plugin_starter.dart';
+export 'src/utils/storable_hash_map.dart';
+export 'src/utils/storable_string.dart';
