+++ lib/src/exceptions/communication_exception.dart
