+++ lib/src/exceptions/malformed_url_exception.dart
