+++ lib/src/message/message.dart
