+++ lib/src/message/message_type.dart
