+++ lib/src/plugin/communication_secret.dart
