+++ lib/src/plugin/declaration.dart
