+++ lib/src/plugin/menu_registrar.dart
