+++ lib/src/plugin/plugin_information.dart
