+++ lib/src/plugin/plugin_information_factory.dart
