+++ lib/src/plugin/plugin_listener.dart
@@ -1,7 +1,7 @@
 library geiger_api;
 
-import 'geiger_url.dart';
-import 'message.dart';
+import '../message/geiger_url.dart';
+import '../message/message.dart';
 
 /// Interface for plugins listening for events.
 abstract class PluginListener {
