+++ lib/src/plugin/plugin_registrar.dart
