+++ lib/src/plugin/plugin_starter.dart
