+++ lib/src/utils/storable_string.dart
