+++ test/api_call_sequences_test.dart
@@ -2,8 +2,8 @@ import 'dart:convert';
 
 import 'package:geiger_api/geiger_api.dart';
 import 'package:geiger_api/src/communication/communication_helper.dart';
-import 'package:geiger_api/src/communication/communication_secret.dart';
-import 'package:geiger_api/src/communication/passthrough_controller.dart';
+import 'package:geiger_api/src/plugin/communication_secret.dart';
+import 'package:geiger_api/src/storage/passthrough_controller.dart';
 import 'package:geiger_localstorage/geiger_localstorage.dart';
 import 'package:test/test.dart';
