+++ test/communication_secret_test.dart
@@ -1,7 +1,7 @@
 import 'dart:math';
 
 import 'package:collection/collection.dart';
-import 'package:geiger_api/src/communication/communication_secret.dart';
+import 'package:geiger_api/src/plugin/communication_secret.dart';
 import 'package:geiger_localstorage/geiger_localstorage.dart';
 import 'package:test/test.dart';
